+++ web/src/routes/search-index.json/+server.ts
@@ -1,0 +1,9 @@
+import { json } from '@sveltejs/kit';
+import { getContent } from '$lib/content/loader/index.js';
+import { createSearchIndex } from '$lib/content/search.js';
+
+export const prerender = true;
+
+export function GET() {
+	return json(createSearchIndex(getContent()));
+}